% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrna-immuno.R
\name{cf_circrna_pipeline}
\alias{cf_circrna_pipeline}
\title{Run full circRNA pipeline}
\usage{
cf_circrna_pipeline(sequence, gene_expression = list())
}
\arguments{
\item{sequence}{RNA sequence string}

\item{gene_expression}{Named list of gene expression values}
}
\value{
List with immune_scores, composite_scores, recommendations, uncertainty
}
\description{
Run full circRNA pipeline
}
