% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-predict.R
\name{cf_epitope_predict}
\alias{cf_epitope_predict}
\title{Predict epitope binding for a single sequence}
\usage{
cf_epitope_predict(bundle_path, sequence, env_params = list())
}
\arguments{
\item{bundle_path}{Path to a trained .joblib EpitopeModelBundle}

\item{sequence}{Peptide amino acid sequence}

\item{env_params}{Named list of environmental parameters}
}
\value{
Numeric scalar (predicted binding)
}
\description{
Predict epitope binding for a single sequence
}
