% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mamba3-encode.R
\name{cf_mamba3_encode}
\alias{cf_mamba3_encode}
\title{Encode sequence with Mamba3Lite encoder}
\usage{
cf_mamba3_encode(sequence)
}
\arguments{
\item{sequence}{Amino acid or nucleotide sequence string}
}
\value{
Named list of numeric vectors (summary, local_pool, meso_pool, global_pool)
}
\description{
Encode sequence with Mamba3Lite encoder
}
