% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhc-encode.R
\name{cf_mhc_detect_class}
\alias{cf_mhc_detect_class}
\title{Detect MHC class from allele string}
\usage{
cf_mhc_detect_class(allele)
}
\arguments{
\item{allele}{MHC allele string}
}
\value{
Character: "I" or "II"
}
\description{
Detect MHC class from allele string
}
