% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhc-encode.R
\name{cf_mhc_encode}
\alias{cf_mhc_encode}
\title{Encode MHC features for a peptide-allele pair}
\usage{
cf_mhc_encode(peptide, allele)
}
\arguments{
\item{peptide}{Peptide sequence string}

\item{allele}{MHC allele string (e.g., "HLA-A*02:01")}
}
\value{
Numeric vector (1018-dim for MHC-I, 947-dim for MHC-II)
}
\description{
Encode MHC features for a peptide-allele pair
}
