% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctm-simulate.R
\name{cf_rna_ctm_params}
\alias{cf_rna_ctm_params}
\title{Derive RNA-CTM parameters from modification and delivery config}
\usage{
cf_rna_ctm_params(
  modification = "none",
  delivery_vector = "LNP_standard",
  route = "IV",
  ires_score = 0.5,
  gc_content = 0.5,
  struct_stability = 0.5,
  innate_immune_score = 0
)
}
\arguments{
\item{modification}{Nucleotide modification type: "none", "m6A", "pseudouridine", "5mC", "ms2m6A"}

\item{delivery_vector}{Delivery vector type: "LNP_standard", "LNP_liver", "AAV", "naked"}

\item{route}{Administration route: "IV", "SC", "IM"}

\item{ires_score}{IRES efficiency score (0-1)}

\item{gc_content}{GC content (0-1)}

\item{struct_stability}{Structure stability score (0-1)}

\item{innate_immune_score}{Innate immune activation score (0-1)}
}
\value{
Named numeric vector with RNA-CTM parameters
}
\description{
Derive RNA-CTM parameters from modification and delivery config
}
