% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-bridge.R
\name{cf_use_python}
\alias{cf_use_python}
\title{Specify Python interpreter for Confluencia}
\usage{
cf_use_python(python_path)
}
\arguments{
\item{python_path}{Path to the Python executable}
}
\description{
Set the Python interpreter to use for the reticulate bridge.
Must be called before any other cf_* function.
}
